MMemperbaiki bug pada detail
