+++ client/src/pages/DetailPhotos.js
@@ -8,9 +8,22 @@ import RandomPhotos from "../components/RandomPhotos";
 function DetailPhotos(props) {
   const [photos, setData] = useState([]);
   const id = props.match.params.id;
-  const [detailPhotos, setDetailPhotos] = useState();
+  const [detailPhotos, setDetailPhotos] = useState(id);
+  const [isLoading, setLoading] = useState(true);
 
-  const getData = () => {
+  const getDownloadFromApi = async () => {
+    const { data } = await axios.get(
+      detailPhotos.data.download +
+        "?client_id=qN-U_v7VlbUf0Yb_91yXwDtXhPgtf3j9LDrzQsWvAww"
+    );
+    const response = await axios.get(data.url, { responseType: "blob" });
+    const url = URL.createObjectURL(new Blob([response.data]));
+    detailPhotos.data.download = url;
+    console.log(url);
+    setLoading(false);
+  };
+
+  const getData = async () => {
     axios
       .get(`http://localhost:5000/photos`)
       .then(response => response.data)
@@ -19,8 +32,8 @@ function DetailPhotos(props) {
       });
   };
 
-  const getDetailPhoto = () => {
-    axios
+  const getDetailPhoto = async () => {
+    await axios
       .get(`http://localhost:5000/detailpage/${id}`)
       .then(response => response.data)
       .then(json => {
@@ -40,9 +53,24 @@ function DetailPhotos(props) {
 
   useEffect(() => {
     getDetailPhoto();
-  }, []);
+  }, [id]);
 
-  return (
+  useEffect(() => {
+    if (detailPhotos.data) {
+      setLoading(true);
+      if (id[0] === "a") {
+        getDownloadFromApi();
+      } else {
+        if (isLoading) {
+          setLoading(false);
+        }
+      }
+    }
+  }, [detailPhotos]);
+
+  return isLoading ? (
+    "loading"
+  ) : (
     <div>
       <Navbar />
       <DetailPhoto {...{ ...detailPhotos, id }} />
