+++ server/middlewares/handlerError.js
@@ -1,5 +1,4 @@
 const handlerError = (err, req, res, next) => {
-  console.log(err);
   if (typeof err.code === "number" && err.code <= 500 && err.message) {
     res.status(err.code).json({
       message: err.message
